pyproject.toml
[project]
name = "discord-bot-project"
version = "0.1.0"
description = "A Discord bot with automoderation and join roles functionalities."
authors = [
    {name = "Your Name", email = "youremail@example.com"}
]
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.0.0",
    "flask>=2.0.0",
    "pyyaml>=6.0",
    "pytest>=7.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "oauthlib>=3.2.2",
    "flask-discord>=0.1.7",
    "flask-limiter>=3.3.1",
    "flask-talisman>=1.0.0",
    "flask-seasurf>=1.1.1",
    "pyopenssl>=23.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.uv.sources]
discord-bot-project = { workspace = true }

[dependency-groups]
dev = [
    "discord-bot-project",
]
